Users can customize discardable errors
Add hashes to `discardable_errors` array on `Sauce::Config` to choose
what to ignore when there are errors quitting sessions.

`discardable_errors` hashes require an `:exception` class and can also
match an optional `:message`.

`:exception` is a class (not a classname).

`:message` is matched with `match` and can thus be a string or Regexp.

Errors are only discarded if thrown when calling `quit` on a session; At
all other times they will be raised.

Example Usage:
```
Sauce.config do |c|
  # The Rest Of Your Config
  c[:discardable_errors] << {:exception => Selenium::WebDriver::Error::UnknownError, :message => "already finished"}
end
```